Move start coords components into interactive-graph-editor dir
