+++ packages/perseus-editor/src/widgets/interactive-graph-editor/start-coords-settings.tsx
@@ -17,7 +17,7 @@ import {spacing} from "@khanacademy/wonder-blocks-tokens";
 import arrowCounterClockwise from "@phosphor-icons/core/bold/arrow-counter-clockwise-bold.svg";
 import * as React from "react";
 
-import Heading from "./heading";
+import Heading from "../../components/heading";
 import StartCoordsAngle from "./start-coords-angle";
 import StartCoordsCircle from "./start-coords-circle";
 import StartCoordsLine from "./start-coords-line";
@@ -25,7 +25,7 @@ import StartCoordsMultiline from "./start-coords-multiline";
 import StartCoordsPoint from "./start-coords-point";
 import StartCoordsQuadratic from "./start-coords-quadratic";
 import StartCoordsSinusoid from "./start-coords-sinusoid";
-import {getDefaultGraphStartCoords} from "./util";
+import {getDefaultGraphStartCoords} from "../../components/util";
 
 import type {PerseusGraphType, Range} from "@khanacademy/perseus";
